backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Vanna AI
vanna>=0.1.0

# Google BigQuery
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0

# Google Generative AI for Gemini integration
google-generativeai>=0.3.0

# Vector database
# chromadb>=0.4.0

# MongoDB for vector storage
pymongo>=4.0.0
sentence-transformers>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# CLI utilities
rich>=13.0.0

# db-dtype 
db-dtypes